+++ OSLC4Net_SDK/OSLC4Net.Core/Exceptions/OslcCoreApplicationException.cs
@@ -20,6 +20,16 @@ namespace OSLC4Net.Core.Exceptions;
 /// </summary>
 public class OslcCoreApplicationException : Exception
 {
+    public OslcCoreApplicationException(string message) :
+        base(message)
+    {
+    }
+
+    public OslcCoreApplicationException(string message, Exception innerException) :
+        base(message, innerException)
+    {
+    }
+
     public OslcCoreApplicationException(string messageKey, object?[] args) :
         base(MessageExtractor.GetMessage(messageKey, args))
     {