+++ OSLC4Net_SDK/OSLC4Net.Domains.RequirementsManagement/RequirementCollection.cs
@@ -17,15 +17,16 @@ public record RequirementCollection : AbstractResourceRecord
     {
     }
 
-
-    [OslcDescription("The resource type URIs.")] // Escape quotes in description
-    [OslcOccurs(Occurs.ZeroOrMany)]
-    [OslcPropertyDefinition("http://www.w3.org/1999/02/22-rdf-syntax-ns#type")]
-    [OslcName("type")]
-    [OslcRepresentation(Representation.Reference)]
-    [OslcReadOnly(false)] // Assuming read_only property exists
-    [OslcTitle("type")] // Use prop name as fallback title
-    public HashSet<Uri> Type { get; set; }
+    //
+    //
+    // [OslcDescription("The resource type URIs.")] // Escape quotes in description
+    // [OslcOccurs(Occurs.ZeroOrMany)]
+    // [OslcPropertyDefinition("http://www.w3.org/1999/02/22-rdf-syntax-ns#type")]
+    // [OslcName("type")]
+    // [OslcRepresentation(Representation.Reference)]
+    // [OslcReadOnly(false)] // Assuming read_only property exists
+    // [OslcTitle("type")] // Use prop name as fallback title
+    [Obsolete] public HashSet<Uri> Type => new(Types);
 
 
     [OslcDescription(